# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/conductor/v1/api",
    "paths": {
        "/contas": {
            "get": {
                "description": "Retorna a lista com todas as contas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Obter contas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ContaResponse"
                            }
                        }
                    }
                }
            }
        },
        "/contas/{id}": {
            "get": {
                "description": "Retorna a lista com todas as contas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Obter contas por id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da conta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ContaResponse"
                            }
                        }
                    }
                }
            }
        },
        "/contas/{id}/transacoes": {
            "get": {
                "description": "Retorna a lista com todas as transações da conta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Obter transações da conta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da conta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.TransacaoResponse"
                            }
                        }
                    }
                }
            }
        },
        "/contas/{id}/transacoes.pdf": {
            "get": {
                "description": "Download de relatório com todas as transações da conta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "summary": "Obter contas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da conta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ContaResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.TransacaoResponse": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_conta": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        }
    }
}